* Fixed issue with verbatim (and figure...) tag types incorrectly displaying extra views in tag editor * Rolled back changes to int/NSInteger for SCHelpers.m - this broke selection. Commented area as a warning (and to-do) for future * Fixed tag preview generator to emit preview text for verbatim * Fixed tag editor to render instructional text for verbatim

+++ StatTag/View Controllers/ScintillaHelpers/SCHelpers.m
@@ -12,7 +12,9 @@
 
 @implementation SCHelpers
 
-+(int)Clamp:(int)value min:(int)min max:(int) max
+//FIXME: fix types - but not now
+//for now, leave these types as they are - there's an issue where fixing this to return types breaks selection - it needs to be fixed, but you need to test it. Last time I "fixed" the types it broke all user click event handling.
++(int)Clamp:(NSInteger)value min:(int)min max:(int) max
 {
   if (value < min)
     return min;
@@ -20,7 +22,7 @@
   if (value > max)// && max > min)//eww changed to add && max > min
     return max;
   
-  return (int)value;
+  return value;
 }
